backups/20260109/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
celery>=5.3.6
redis>=5.0.1
google-generativeai>=0.3.2
google-genai
# 影片生成相關
Pillow>=10.0.0
edge-tts>=6.1.0